:recycle: refactoring: remove duplicated codes and fix frame issue
bust샷만 생성되던 프롬프트 이슈 해결 및 중복 코드 base_generator로 분리
